// src/app/models/lead.ts
export class Lead extends BaseModel {
    id: string;
    dateCreated: Date;
    dateModified: Date;
    modifiedUserId: string;
    createdBy: string;
    description: string;
    deleted: boolean;
    salesRepId: string;
    salesRep: string;
    salutation: string;
    firstName: string;
    lastName: string;
    title: string;
    department: string;
    email: string;
    doNotCall: boolean;
    phoneHome: string;
    phoneMobile: string;
    phoneWork: string;
    phoneOther: string;
    phoneFax: string;
    primaryAddressStreet: string;
    primaryAddressStreet2: string;
    primaryAddressCity: string;
    primaryAddressState: string;
    primaryAddressPostalcode: string;
    primaryAddressCountry: string;
    altAddressStreet: string;
    altAddressStreet2: string;
    altAddressCity: string;
    altAddressState: string;
    altAddressPostalcode: string;
    altAddressCountry: string;
    assistant: string;
    assistantPhone: string;
    converted: boolean;
    referedBy: string;
    leadSource: string;
    leadSourceDescription: string;
    status: string;
    statusDescription: string;
    reportsToId: string;
    accountDescription: string;
    contactId: string;
    contactName: string;
    accountId: string;
    accountName: string;
    opportunityId: string;
    opportunityName: string;
    opportunityAmount: string;
    campaignId: string;
    birthdate: string;
    portalName: string;
    portalApp: string;
    website: string;
    corporateIdentityId: string;
    rating: string;
    channel: any;
    merchandiseInterest : any;
    contactType: any;
    setData(data) {
        this.setString(data, 'id');
        this.setDate(data, 'dateCreated');
        this.setDate(data, 'dateModified');
        this.setString(data, 'createdBy');
        this.setString(data, 'description');
        this.setBoolean(data, 'deleted');
        this.setString(data, 'salesRepId');
        this.setString(data, 'salesRep');
        this.setString(data, 'salutation');
        this.setString(data, 'firstName');
        this.setString(data, 'lastName');
        this.setString(data, 'title');
        this.setString(data, 'department');
        this.setString(data, 'email');
        this.setBoolean(data, 'doNotCall');
        this.setString(data, 'doNotCall');
        this.setString(data, 'phoneMobile');
        this.setString(data, 'phoneWork');
        this.setString(data, 'phoneOther');
        this.setString(data, 'phoneFax');
        this.setString(data, 'primaryAddressStreet');
        this.setString(data, 'primaryAddressStreet2');
        this.setString(data, 'primaryAddressCity');
        this.setString(data, 'primaryAddressState');
        this.setString(data, 'primaryAddressPostalcode');
        this.setString(data, 'primaryAddressCountry');
        this.setString(data, 'altAddressStreet');
        this.setString(data, 'altAddressStreet2');
        this.setString(data, 'altAddressCity');
        this.setString(data, 'altAddressState');
        this.setString(data, 'altAddressPostalcode');
        this.setString(data, 'altAddressCountry');
        this.setString(data, 'assistant');
        this.setString(data, 'assistantPhone');
        this.setBoolean(data, 'converted');
        this.setString(data, 'referedBy');
        this.setString(data, 'leadSource');
        this.setString(data, 'leadSourceDescription');
        this.setString(data, 'status');
        this.setString(data, 'statusDescription');
        this.setString(data, 'reportsToId');
        this.setString(data, 'accountDescription');
        this.setString(data, 'contactId');
        this.setString(data, 'contactName');
        this.setString(data, 'accountId');
        this.setString(data, 'accountName');
        this.setString(data, 'opportunityId');
        this.setString(data, 'opportunityName');
        this.setString(data, 'opportunityAmount');
        this.setString(data, 'campaignId');
        this.setString(data, 'birthdate');
        this.setString(data, 'portalName');
        this.setString(data, 'portalApp');
        this.setString(data, 'website');
        this.setString(data, 'corporateIdentityId');
        this.setString(data, 'rating');
        this.setArray(data, 'channel');
        this.setArray(data, 'merchandiseInterest');
        this.setArray(data, 'contactType');
    }
}